semantic-kernel/requirements.txt
semantic_kernel[azure] == 1.34.0
python-dotenv